Small cosmetic fixes for instances and a couple of typo fixes (#792)
* Fix browser tab saying "- Ruqqus"

The browser tab says "Create a post - Ruqqus" even if you have changed the name of the site in your instance, so this fixes that.

* board_banned.html instance tab title fix

* Typo fix in conduct.html

* Replaced "Ruqqus" in delete_comment_modal.html

Replaced "Ruqqus" with environment variable.

* Replaced "Ruqqus" in delete_post_modal.html

* Replaced "Ruqqus" in flag_comment_modal.html

* Replaced "Ruqqus" in flag_post_modal.html

* Replaced "Ruqqus" in help.html

I don't know if you can do this on the meta tag, so I'm leaving that be just in case.

* Replaced "Ruqqus" in make_board.html

* Replaced "Ruqqus" in tab title in userpage.html

* Typo fixes in login.py welcome message

Doing the "{{ 'SITE_NAME' | app_config }}" thing in the message didn't work. Maybe should find a fix for that.